feat: move rating to escrow and add it to user metadata
