+++ app/js/features/escrow/actions.js
@@ -1,6 +1,7 @@
 import {
   CREATE_ESCROW, RESET_CREATE_ESCROW_STATUS, LOAD_ESCROWS, RELEASE_ESCROW, CANCEL_ESCROW,
-  RATE_TRANSACTION, PAY_ESCROW, OPEN_CASE, OPEN_CASE_SIGNATURE, PAY_ESCROW_SIGNATURE, CLOSE_DIALOG
+  RATE_TRANSACTION, PAY_ESCROW, OPEN_CASE, OPEN_CASE_SIGNATURE, PAY_ESCROW_SIGNATURE, CLOSE_DIALOG,
+  ADD_USER_RATING, USER_RATING
 } from './constants';
 import Escrow from 'Embark/contracts/Escrow';
 import { addDecimals } from '../../utils/numbers';
@@ -36,3 +37,7 @@ export const cancelEscrow = (escrowId) => ({ type: CANCEL_ESCROW, escrowId, toSe
 export const rateTransaction = (escrowId, rating) => ({ type: RATE_TRANSACTION, escrowId, rating, toSend: Escrow.methods.rateTransaction(escrowId, rating) });
 
 export const closeDialog = () => ({ type: CLOSE_DIALOG });
+
+export const checkUserRating = (address) => ({ type: USER_RATING, address });
+
+export const addUserRating = () => ({ type: ADD_USER_RATING });
