+++ app/js/features/escrow/constants.js
@@ -49,3 +49,8 @@ export const GET_ARBITRATION_BY_ID_SUCCEEDED = 'GET_ARBITRATION_BY_ID_SUCCEEDED'
 export const GET_ARBITRATION_BY_ID_FAILED = 'GET_ARBITRATION_BY_ID_FAILED';
 
 export const CLOSE_DIALOG = 'CLOSE_DIALOG';
+
+export const USER_RATING = 'CHECK_USER_RATING';
+export const ADD_USER_RATING = 'ADD_USER_RATING';
+export const USER_RATING_SUCCEEDED = 'CHECK_USER_RATING_SUCCEEDED';
+export const USER_RATING_FAILED = 'CHECK_USER_RATING_FAILED';
