+++ app/js/features/license/constants.js
@@ -14,9 +14,3 @@ export const CHECK_LICENSE_OWNER_FAILED = 'CHECK_LICENSE_OWNER_FAILED';
 export const GET_LICENSE_OWNERS = 'GET_LICENSE_OWNERS';
 export const GET_LICENSE_OWNERS_SUCCCEDED = 'GET_LICENSE_OWNERS_SUCCCEDED';
 export const GET_LICENSE_OWNERS_FAILED = 'GET_LICENSE_OWNERS_FAILED';
-
-export const USER_RATING = 'CHECK_USER_RATING';
-export const ADD_USER_RATING = 'ADD_USER_RATING';
-export const USER_RATING_SUCCEEDED = 'CHECK_USER_RATING_SUCCEEDED';
-export const USER_RATING_FAILED = 'CHECK_USER_RATING_FAILED';
-
